# packages/core/docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stellar Lense API",
    "description": "Stellar Lense REST API for Stellar wallet risk scoring, alert management, AMM pool risk, cross-chain linking, and compliance exports. See [ReDoc](/redoc) for an alternative documentation view.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/stream/scores": {
      "get": {
        "tags": [
          "Score Streaming (SSE)"
        ],
        "summary": "Subscribe to real-time score updates via SSE",
        "description": "Server-Sent Events endpoint. Provide comma-separated Stellar wallet addresses (max 50) to subscribe to score updates. On reconnect, send the `Last-Event-ID` header to replay missed events from the last 5 minutes.",
        "operationId": "stream_scores_stream_scores_get",
        "parameters": [
          {
            "name": "wallets",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Comma-separated Stellar wallet addresses to subscribe to (max 50 chars per address, max 50 wallets per connection).",
              "title": "Wallets"
            },
            "description": "Comma-separated Stellar wallet addresses to subscribe to (max 50 chars per address, max 50 wallets per connection)."
          },
          {
            "name": "Last-Event-ID",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last received SSE event ID for reconnect replay.",
              "title": "Last-Event-Id"
            },
            "description": "Last received SSE event ID for reconnect replay."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Invalid wallet address format."
          },
          "429": {
            "description": "Connection limit exceeded for this API key."
          }
        }
      }
    },
...
    "/v1/scores": {
      "get": {
        "summary": "List Scores",
        "description": "Return latest scores filtered by score, flags, paging, and `sort_by` ordering.",
        "operationId": "list_scores_v1_scores_get",
        "parameters": [
          {
            "name": "min_score",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 0,
              "description": "Minimum RiskScore (0-100).",
              "default": 0,
              "title": "Min Score"
            },
            "description": "Minimum RiskScore (0-100)."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "benford_flag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by the Benford detector flag when provided.",
              "title": "Benford Flag"
            },
            "description": "Filter by the Benford detector flag when provided."
          },
          {
            "name": "ml_flag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by the machine-learning detector flag when provided.",
              "title": "Ml Flag"
            },
            "description": "Filter by the machine-learning detector flag when provided."
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(score|confidence|timestamp)$",
              "description": "Sort descending by score, confidence, or timestamp.",
              "default": "score",
              "title": "Sort By"
            },
            "description": "Sort descending by score, confidence, or timestamp."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RiskScore"
                  },
                  "title": "Response List Scores V1 Scores Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RiskScore": {
        "properties": {
          "wallet": {
            "type": "string",
            "title": "Wallet"
          },
          "asset_pair": {
            "type": "string",
            "title": "Asset Pair"
          },
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score",
            "description": "0-100; higher = more suspicious"
          },
          "benford_flag": {
            "type": "boolean",
            "title": "Benford Flag"
          },
          "ml_flag": {
            "type": "boolean",
            "title": "Ml Flag"
          },
          "confidence": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "disputed": {
            "type": "boolean",
            "title": "Disputed",
            "default": false
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "latency_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ms",
            "description": "End-to-end latency in milliseconds from trade receipt to score update"
          },
          "score_lower": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Lower",
            "description": "Lower bound of 90 % conformal prediction interval"
          },
          "score_upper": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Upper",
            "description": "Upper bound of 90 % conformal prediction interval"
          },
          "prediction_set": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prediction Set",
            "description": "Class indices in the conformal prediction set"
          },
          "coverage_guarantee": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverage Guarantee",
            "description": "Target coverage level (1 - alpha) of the prediction set"
          }
        },
        "type": "object",
        "required": [
          "wallet",
          "asset_pair",
          "score",
          "benford_flag",
          "ml_flag",
          "confidence",
          "timestamp"
        ],
        "title": "RiskScore"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  },
...
}